+++ CobaltTests/CobaltTests-Authentication.swift
@@ -8,13 +8,43 @@
 
 
 import XCTest
-import Nimble
 import Alamofire
 import Combine
 import Foundation
 @testable import Cobalt
 
 class CobaltTestsAuthentication: CobaltTests {
+    override func setUp() {
+        super.setUp()
+        config.authentication.authorization = .basicHeader
+        config.host = "https://apps.e-sites.nl"
+        config.authentication.path = "/prototypes/cobalt/access_token.json"
+    }
+    
+    func testsAuthenticate() {
+        waitUntil { done in
+            let request = Request {
+                $0.authentication = .oauth2(.clientCredentials)
+                $0.path = "/prototypes/cobalt/users.php"
+            }
+
+            self.client.request(request)
+                .sink(receiveCompletion: { event in
+                    switch event {
+                    case .finished:
+                        break
+                    case .failure(let error):
+                        XCTAssert(false, "\(error)")
+                    }
+                    done?()
+                }, receiveValue: { response in
+                    guard let dic = response as? [String: Any], let users = dic["users"] as? [Any] else {
+                        return
+                    }
+                    XCTAssert(users.count == 2)
+                }).store(in: &self.cancellables)
+        }
+    }
 
     func testDoubleAccessTokens() {
         let host1 = "https://www.e-sites.nl"
@@ -32,8 +62,7 @@ class CobaltTestsAuthentication: CobaltTests {
         accessToken2.expireDate = Date(timeIntervalSinceNow: 20)
         accessToken2.grantType = .password
         accessToken2.store()
-
-        expect(AccessToken.get(host: host1, grantType: .password)).to(beNil())
+        XCTAssert(AccessToken.get(host: host1, grantType: .password) == nil)
 
         guard let getAccessToken1 = AccessToken.get(host: host1, grantType: .clientCredentials) else {
             XCTAssert(false, "Should have gotten accessToken1")
@@ -45,21 +74,21 @@ class CobaltTestsAuthentication: CobaltTests {
             return
         }
 
-        expect(accessToken1.accessToken) == getAccessToken1.accessToken
-        expect(getAccessToken1.refreshToken).to(beNil())
-        expect(accessToken1.refreshToken).to(beNil())
-        expect(accessToken1.grantType) == getAccessToken1.grantType
-        expect(accessToken1.expireDate?.timeIntervalSince1970) == getAccessToken1.expireDate?.timeIntervalSince1970
+        XCTAssert(accessToken1.accessToken == getAccessToken1.accessToken)
+        XCTAssert(getAccessToken1.refreshToken == nil)
+        XCTAssert(accessToken1.refreshToken == nil)
+        XCTAssert(accessToken1.grantType == getAccessToken1.grantType)
+        XCTAssert(accessToken1.expireDate?.timeIntervalSince1970 == getAccessToken1.expireDate?.timeIntervalSince1970)
 
-        expect(accessToken2.accessToken) == getAccessToken2.accessToken
-        expect(accessToken2.refreshToken) == getAccessToken2.refreshToken
-        expect(accessToken2.grantType) == getAccessToken2.grantType
-        expect(accessToken2.expireDate?.timeIntervalSince1970) == getAccessToken2.expireDate?.timeIntervalSince1970
+        XCTAssert(accessToken2.accessToken == getAccessToken2.accessToken)
+        XCTAssert(accessToken2.refreshToken == getAccessToken2.refreshToken)
+        XCTAssert(accessToken2.grantType == getAccessToken2.grantType)
+        XCTAssert(accessToken2.expireDate?.timeIntervalSince1970 == getAccessToken2.expireDate?.timeIntervalSince1970)
 
         accessToken1.clear()
-        expect(AccessToken.get(host: host1, grantType: .clientCredentials)).to(beNil())
+        XCTAssert(AccessToken.get(host: host1, grantType: .clientCredentials) == nil)
 
         accessToken2.clear()
-        expect(AccessToken.get(host: host2, grantType: .password)).to(beNil())
+        XCTAssert(AccessToken.get(host: host2, grantType: .password) == nil)
     }
 }