+++ CobaltTests/CobaltTests-Cache.swift
@@ -39,8 +39,8 @@ class CobaltTestsCache: CobaltTests {
                     break
                 case .failure(let error):
                     XCTAssert(false, "\(error)")
-                    done()
                 }
+                done?()
             }, receiveValue: { response in
                 if let dictionary = response as? [String: Any], let data = dictionary["data"] as? [Any] {
                     expect(data.count) == 10
@@ -65,7 +65,7 @@ class CobaltTestsCache: CobaltTests {
                         case .failure(let error):
                             XCTAssert(false, "\(error)")
                         }
-                        done()
+                        done?()
                     }, receiveValue: { response in
                         if let dictionary = response as? [String: Any], let data = dictionary["data"] as? [Any] {
                             expect(data.count) == 10
@@ -95,7 +95,7 @@ class CobaltTestsCache: CobaltTests {
                     break
                 case .failure(let error):
                     XCTAssert(false, "\(error)")
-                    done()
+                    done?()
                 }
             }, receiveValue: { response in
                 if let dictionary = response as? [String: Any], let data = dictionary["data"] as? [Any] {
@@ -120,7 +120,7 @@ class CobaltTestsCache: CobaltTests {
                     case .failure(let error):
                         XCTAssert(false, "\(error)")
                     }
-                    done()
+                    done?()
                 }, receiveValue: { resonse2 in
                     if let dictionary = resonse2 as? [String: Any], let data = dictionary["data"] as? [Any] {
                         expect(data.count) == 5