+++ Sources/Core/Extensions/String+Extension.swift
@@ -16,4 +16,26 @@ extension String {
             String(format: "%02hhx", $0)
         }.joined()
     }
+    
+    var urlEncoded: String {
+        return self.addingPercentEncoding(withAllowedCharacters: .urlHostAllowed) ?? self
+    }
+    
+    /// Combine the host and path
+    /// - Parameters:
+    ///   - host: `String`
+    ///   - path: `String`
+    /// - Returns: `String`
+    static func combined(host: String, path: String) -> String {
+        var newHost = host
+        var newPath = path
+        
+        if host.hasSuffix("/") {
+            newHost = String(host.dropLast())
+        }
+        if path.hasPrefix("/") {
+            newPath = String(path.dropFirst())
+        }
+        return newHost + "/" + newPath
+    }
 }